refactor: DetailScreen을 Stateless 하도록 수정
