+++ app/src/main/java/woowacourse/shopping/presentation/detail/ui/DetailScreen.kt
@@ -5,9 +5,7 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.navigationBarsPadding
@@ -18,55 +16,36 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.viewmodel.compose.viewModel
-import coil3.compose.AsyncImage
-import kotlinx.coroutines.launch
-import org.w3c.dom.Text
 import woowacourse.shopping.R
-import woowacourse.shopping.domain.model.AddItemResult
-import woowacourse.shopping.presentation.common.QuantityCounter
 import woowacourse.shopping.presentation.common.ShoppingAppBar
-import woowacourse.shopping.presentation.detail.viewmodel.DetailViewModel
-import woowacourse.shopping.ui.theme.Gray40
+import woowacourse.shopping.presentation.detail.model.DetailUiState
+import woowacourse.shopping.presentation.detail.ui.components.DetailContent
+import woowacourse.shopping.presentation.detail.ui.components.LastSeenProductCard
 import woowacourse.shopping.ui.theme.Green40
-import woowacourse.shopping.util.formattedPrice
 
 @Composable
 fun DetailScreen(
-    id: Long,
-    isFromLastSeen: Boolean = false,
-    onNavigateToCart: (AddItemResult) -> Unit,
+    uiState: DetailUiState,
     onClickLastProductCard: (Long) -> Unit,
     onBack: () -> Unit,
+    onAddToCart: () -> Unit,
+    onIncrease: () -> Unit,
+    onDecrease: () -> Unit,
     modifier: Modifier = Modifier,
-    viewModel: DetailViewModel = viewModel(),
 ) {
-    val scope = rememberCoroutineScope()
-    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
-    val product = uiState.product
-
-    LaunchedEffect(Unit) {
-        viewModel.loadProduct(id, isFromLastSeen)
-    }
-
     Scaffold(
         containerColor = Color.White,
         topBar = {
@@ -93,16 +72,7 @@ fun DetailScreen(
                         .fillMaxWidth()
                         .height(48.dp)
                         .background(Green40)
-                        .clickable {
-                            scope.launch {
-                                val result =
-                                    viewModel.addToCart(
-                                        id = id,
-                                        quantity = uiState.quantity,
-                                    )
-                                onNavigateToCart(result)
-                            }
-                        },
+                        .clickable { onAddToCart() },
                 contentAlignment = Alignment.Center,
             ) {
                 Text(
@@ -122,16 +92,12 @@ fun DetailScreen(
             verticalArrangement = Arrangement.spacedBy(30.dp),
         ) {
             DetailContent(
-                imageUrl = product.imageUrl,
-                productName = product.name,
+                imageUrl = uiState.product.imageUrl,
+                productName = uiState.product.name,
                 price = uiState.price,
                 quantity = uiState.quantity,
-                onIncrease = {
-                    scope.launch { viewModel.increase() }
-                },
-                onDecrease = {
-                    scope.launch { viewModel.decrease() }
-                },
+                onIncrease = { onIncrease() },
+                onDecrease = { onDecrease() },
                 modifier = Modifier.padding(innerPadding),
             )
             if (uiState.showLastSeenProductCard) {
@@ -145,87 +111,15 @@ fun DetailScreen(
     }
 }
 
-@Composable
-private fun DetailContent(
-    imageUrl: String,
-    productName: String,
-    price: Long,
-    quantity: Int,
-    onIncrease: () -> Unit,
-    onDecrease: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(16.dp),
-    ) {
-        AsyncImage(
-            model = imageUrl,
-            contentDescription = productName,
-            contentScale = ContentScale.Crop,
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .aspectRatio(1f),
-        )
-        Text(
-            text = productName,
-            fontSize = 24.sp,
-            fontWeight = FontWeight.Bold,
-            color = Color.Black,
-            modifier = Modifier.padding(horizontal = 18.dp),
-        )
-        HorizontalDivider(modifier = Modifier.fillMaxWidth(), color = Gray40)
-        Row(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 18.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween,
-        ) {
-            Text(
-                text = formattedPrice(price),
-                fontSize = 20.sp,
-                fontWeight = FontWeight.Normal,
-                lineHeight = 26.sp,
-                color = Color.Black,
-                modifier =
-                    Modifier
-                        .weight(.5f),
-            )
-            QuantityCounter(
-                quantity = quantity,
-                onIncrease = onIncrease,
-                onDecrease = onDecrease,
-                modifier =
-                    Modifier
-                        .weight(.5f),
-            )
-        }
-    }
-}
-
 @Preview
 @Composable
 private fun DetailScreenPreview() {
     DetailScreen(
-        id = 1L,
-        onNavigateToCart = {},
+        uiState = DetailUiState(),
         onClickLastProductCard = {},
         onBack = {},
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun DetailContentPreview() {
-    DetailContent(
-        imageUrl = "",
-        productName = "Test",
-        price = 10000,
-        onDecrease = {},
+        onAddToCart = {},
         onIncrease = {},
-        quantity = 3,
+        onDecrease = {},
     )
 }
